support for heif and hopefully mpo image formats by converting to jpeg
